ADD :: 열고 닫기 기능
