+++ src/types/common/header.type.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ProfilePopoverPropsType {
+  isOpen: boolean;
+  close: Dispatch<SetStateAction<boolean>>;
+}
